update check file file
