--- tasks/maintenance/test.py
@@ -1,4 +1,0 @@
-import pywikibot
-site = pywikibot.Site()
-page = pywikibot.Page(site,"عقيد_(رتبة_عسكرية)")
-print(page.text)
